+++ src/main/java/net/giok3r/whereami/Config.java
@@ -15,49 +15,34 @@ import net.neoforged.neoforge.common.ModConfigSpec;
 // An example config class. This is not required, but it's a good idea to have one to keep your config organized.
 // Demonstrates how to use Neo's config APIs
 @EventBusSubscriber(modid = WhereAmI.MODID)
-public class Config
-{
+public class Config {
     private static final ModConfigSpec.Builder BUILDER = new ModConfigSpec.Builder();
 
-    private static final ModConfigSpec.BooleanValue LOG_DIRT_BLOCK = BUILDER
-            .comment("Whether to log the dirt block on common setup")
-            .define("logDirtBlock", true);
-
-    private static final ModConfigSpec.IntValue MAGIC_NUMBER = BUILDER
-            .comment("A magic number")
-            .defineInRange("magicNumber", 42, 0, Integer.MAX_VALUE);
-
-    public static final ModConfigSpec.ConfigValue<String> MAGIC_NUMBER_INTRODUCTION = BUILDER
-            .comment("What you want the introduction message to be for the magic number")
-            .define("magicNumberIntroduction", "The magic number is... ");
-
-    // a list of strings that are treated as resource locations for items
-    private static final ModConfigSpec.ConfigValue<List<? extends String>> ITEM_STRINGS = BUILDER
-            .comment("A list of items to log on common setup.")
-            .defineListAllowEmpty("items", List.of("minecraft:iron_ingot"), Config::validateItemName);
+    private static final ModConfigSpec.BooleanValue DISPLAY_LOCATION = BUILDER
+            .comment("Show the player's location on screen")
+            .define("display.location", true);
+    private static final ModConfigSpec.BooleanValue DISPLAY_DIRECTION = BUILDER
+            .comment("Show the compass direction the player is facing next to their coordinates")
+            .define("display.direction", true);
+    private static final ModConfigSpec.BooleanValue DISPLAY_BIOME = BUILDER
+            .comment("Show the biome the player is in")
+            .define("display.biome", true);
+    private static final ModConfigSpec.BooleanValue DISPLAY_TIME_AND_DAY = BUILDER
+            .comment("Show the time and in-game day")
+            .define("display.time_and_day", true);
 
     static final ModConfigSpec SPEC = BUILDER.build();
 
-    public static boolean logDirtBlock;
-    public static int magicNumber;
-    public static String magicNumberIntroduction;
-    public static Set<Item> items;
-
-    private static boolean validateItemName(final Object obj)
-    {
-        return obj instanceof String itemName && BuiltInRegistries.ITEM.containsKey(ResourceLocation.parse(itemName));
-    }
+    public static boolean displayLocation;
+    public static boolean displayDirection;
+    public static boolean displayBiome;
+    public static boolean displayTimeAndDay;
 
     @SubscribeEvent
-    static void onLoad(final ModConfigEvent event)
-    {
-        logDirtBlock = LOG_DIRT_BLOCK.get();
-        magicNumber = MAGIC_NUMBER.get();
-        magicNumberIntroduction = MAGIC_NUMBER_INTRODUCTION.get();
-
-        // convert the list of strings into a set of items
-        items = ITEM_STRINGS.get().stream()
-                .map(itemName -> BuiltInRegistries.ITEM.getValue(ResourceLocation.parse(itemName)))
-                .collect(Collectors.toSet());
+    static void onLoad(final ModConfigEvent event) {
+        displayLocation = DISPLAY_LOCATION.get();
+        displayDirection = DISPLAY_DIRECTION.get();
+        displayBiome = DISPLAY_BIOME.get();
+        displayTimeAndDay = DISPLAY_TIME_AND_DAY.get();
     }
 }