+++ src/main/java/net/giok3r/whereami/WhereAmIClient.java
@@ -1,0 +1,12 @@
+package net.giok3r.whereami;
+
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.fml.ModContainer;
+import net.neoforged.neoforge.client.gui.ConfigurationScreen;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+
+public class WhereAmIClient {
+    public WhereAmIClient(IEventBus modEventBus, ModContainer modContainer) {
+        modContainer.registerExtensionPoint(IConfigScreenFactory.class, ConfigurationScreen::new);
+    }
+}