+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.21.7-1.0.2] - Unreleased
 - Updated to Minecraft 1.21.7 (giok3r)
+### Added
+- Your direction is now shown beside your position (SilentChaos512)
 
 ## [1.21.6-1.0.1] - 2025-06-19
 - Updated to Minecraft 1.21.6