+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.21.7-1.0.2] - Unreleased
-- Updated to Minecraft 1.21.7 (giok3r)
+## [1.21.7-1.0.2] - 2025-07-04
+- Updated to Minecraft 1.21.7
 ### Added
-- Your direction is now shown beside your position (SilentChaos512)
+- Your direction is now shown beside your position
+- Config options to edit what information is shown
+- Configuration screen (in Mods menu)
+### Fixed
+- Information overlay is now hidden when the debug screen is displayed
 
 ## [1.21.6-1.0.1] - 2025-06-19
 - Updated to Minecraft 1.21.6